Move OSD lexer/parser/writer into osd subpackage (issue #41)
Stage 2 of the package-restructuring project: osd_lexer.go, osd_parser.go,
osd_writer.go (and their _test.go files) move into a new osd subpackage
(import path github.com/omnist-dev/omnist-go/osd), with the one deliberate
naming change ReadOSD -> osd.Read / WriteOSD -> osd.Write. Every call site
is updated: cmd/omnist/{cmd_infer,cmd_materialize,cmd_schema,cmd_validate}.go,
tools/conformance/drivers.go, and doc-comment references in oml_writer.go/
algebra_test.go/referee_coverage_test.go.

Since osd now imports the root package for Schema/Record/etc., an internal
(package omnist) test file cannot import osd without an import-cycle error
in the test build. Root-package tests that only needed OSD-text fixtures
plus exported API (extract_test.go, referee_test.go,
referee_coverage_test.go, and most of algebra_test.go/lint_test.go/
compatible_with_test.go/normalize_test.go) moved to the external
"omnist_test" package, sharing mustParseOSD/mustOSD/mustOML from the new
osd_external_test_helpers_test.go. The handful of tests needing unexported
production access (le, scalarSub, reachablePlain, computeLocalSignature,
computeRefineKey, appendEscapedLabel, appendUint) stayed in the internal
package; the six of those that also needed an OSD-text fixture now build
the *Record directly as a struct literal instead, since computeLocalSignature/
computeRefineKey only ever look at rec.Fields.

Verified via a before/after conformance-harness run (byte-for-byte
identical output against vendor/omnist-spec/test-suite) and a baseline
worktree at the parent commit confirming coverage is unchanged (root
100.0%, osd 100.0%, cmd/omnist 99.2%, tools/conformance 14.1% -- all
matching the pre-move numbers). go build/vet/test and golangci-lint all
clean.
