+++ cmd/omnist/cmd_infer.go
@@ -6,6 +6,7 @@ import (
 	"io"
 
 	omnist "github.com/omnist-dev/omnist-go"
+	"github.com/omnist-dev/omnist-go/osd"
 )
 
 // cmdInfer implements `omnist infer`: read one or more sample documents
@@ -67,7 +68,7 @@ func cmdInfer(args []string, stdin io.Reader, stdout, stderr io.Writer) int {
 	for _, fb := range fallbacks {
 		_, _ = fmt.Fprintf(stderr, "%s: opened to any: %s\n", fb.Location, fb.Reason)
 	}
-	if err := writeOutput(*out, omnist.WriteOSD(schema, false), stdout); err != nil {
+	if err := writeOutput(*out, osd.Write(schema, false), stdout); err != nil {
 		_, _ = fmt.Fprintf(stderr, "%s: %v\n", name, err)
 		return ExitUsage
 	}
