+++ osd/osd_parser.go
@@ -1,8 +1,12 @@
-package omnist
+package osd
 
-import "fmt"
+import (
+	"fmt"
 
-// ReadOSD parses OSD source text into a Schema (spec ch.5, "text to
+	omnist "github.com/omnist-dev/omnist-go"
+)
+
+// Read parses OSD source text into a Schema (spec ch.5, "text to
 // Schema"). Unlike ReadOML, there is no Limits parameter: an OSD schema is
 // authored, not attacker-controlled arbitrary data (spec §2.4/§4.7's
 // safety limits exist for Documents built from untrusted input), and
@@ -11,11 +15,11 @@ import "fmt"
 // re-reads, not by nesting depth of adversarial data.
 //
 // On failure the returned error is either:
-//   - a *ParseError with a text-position ("line:col") Path, for a genuine
+//   - a *omnist.ParseError with a text-position ("line:col") Path, for a genuine
 //     syntax failure (an unexpected token, an unterminated string, a
 //     control character) — these happen before enough structure exists to
 //     name a record or field; or
-//   - a Diagnostic with a Schema-shaped Path ("RecordName", "RecordName.label",
+//   - an omnist.Diagnostic with a Schema-shaped Path ("RecordName", "RecordName.label",
 //     or the whole-schema fallback "$"), for a schema.* well-formedness
 //     failure (spec §3.3 S-1..S-7) — these happen once at least a partial
 //     Schema exists to describe the location within.
@@ -23,18 +27,18 @@ import "fmt"
 // This split follows spec §8.4 exactly: "A parse.* diagnostic's path MUST
 // be a text-position path. A ... schema.* ... diagnostic's path MUST be a
 // Document or Schema path ... never a text-position path."
-func ReadOSD(text string) (Schema, error) {
+func Read(text string) (omnist.Schema, error) {
 	p := &osdParser{lex: newOSDLexer(text)}
 	if err := p.advance(); err != nil {
-		return Schema{}, err
+		return omnist.Schema{}, err
 	}
 	return p.parseSchema()
 }
 
-// schemaError constructs the Diagnostic (used as an error) a schema.*
-// well-formedness failure reports, per spec §8.4's Schema path forms.
-func schemaError(path string, code Code, msg string) error {
-	return Diagnostic{Path: path, Code: code, Message: msg, Severity: SeverityError}
+// schemaError constructs the omnist.Diagnostic (used as an error) a schema.*
+// well-formedness failure reports, per spec §8.4's omnist.Schema path forms.
+func schemaError(path string, code omnist.Code, msg string) error {
+	return omnist.Diagnostic{Path: path, Code: code, Message: msg, Severity: omnist.SeverityError}
 }
 
 type osdParser struct {
@@ -42,7 +46,7 @@ type osdParser struct {
 	cur osdToken
 }
 
-func (p *osdParser) advance() *ParseError {
+func (p *osdParser) advance() *omnist.ParseError {
 	tok, err := p.lex.next()
 	if err != nil {
 		return err
@@ -51,8 +55,8 @@ func (p *osdParser) advance() *ParseError {
 	return nil
 }
 
-func (p *osdParser) errAt(t osdToken, code Code, msg string) *ParseError {
-	return &ParseError{Line: t.line, Col: t.col, Path: fmt.Sprintf("%d:%d", t.line, t.col), Code: code, Message: msg}
+func (p *osdParser) errAt(t osdToken, code omnist.Code, msg string) *omnist.ParseError {
+	return &omnist.ParseError{Line: t.line, Col: t.col, Path: fmt.Sprintf("%d:%d", t.line, t.col), Code: code, Message: msg}
 }
 
 // parseSchema implements `schema = *( record-def / root-def )` (spec
@@ -61,30 +65,30 @@ func (p *osdParser) errAt(t osdToken, code Code, msg string) *ParseError {
 // one root, unique record names (checked incrementally as each record-def
 // is parsed, so the duplicate is reported as soon as it is seen), and every
 // reference — including the root — resolving to a defined record.
-func (p *osdParser) parseSchema() (Schema, error) {
-	schema := Schema{Env: map[string]*Record{}}
+func (p *osdParser) parseSchema() (omnist.Schema, error) {
+	schema := omnist.Schema{Env: map[string]*omnist.Record{}}
 	var recordOrder []string
 	rootSet := false
 
 	for p.cur.kind != osdTokEOF {
 		if p.cur.kind != osdTokName {
-			return Schema{}, p.errAt(p.cur, CodeParseUnexpectedToken, "expected 'record' or 'root'")
+			return omnist.Schema{}, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected 'record' or 'root'")
 		}
 		switch p.cur.text {
 		case "record":
 			rec, err := p.parseRecordDef()
 			if err != nil {
-				return Schema{}, err
+				return omnist.Schema{}, err
 			}
 			if _, exists := schema.Env[rec.Name]; exists {
-				return Schema{}, schemaError(rec.Name, CodeSchemaDuplicateRecord, fmt.Sprintf("record %q is already defined", rec.Name))
+				return omnist.Schema{}, schemaError(rec.Name, omnist.CodeSchemaDuplicateRecord, fmt.Sprintf("record %q is already defined", rec.Name))
 			}
 			schema.Env[rec.Name] = rec
 			recordOrder = append(recordOrder, rec.Name)
 		case "root":
 			name, err := p.parseRootDef()
 			if err != nil {
-				return Schema{}, err
+				return omnist.Schema{}, err
 			}
 			// spec §5.8 / chapter 9 divergence-ledger D-2: a second `root`
 			// declaration's behavior is an explicitly acknowledged open
@@ -101,15 +105,15 @@ func (p *osdParser) parseSchema() (Schema, error) {
 				rootSet = true
 			}
 		default:
-			return Schema{}, p.errAt(p.cur, CodeParseUnexpectedToken, "expected 'record' or 'root'")
+			return omnist.Schema{}, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected 'record' or 'root'")
 		}
 	}
 
 	if !rootSet {
-		return Schema{}, schemaError("$", CodeSchemaNoRoot, "a schema must declare a root")
+		return omnist.Schema{}, schemaError("$", omnist.CodeSchemaNoRoot, "a schema must declare a root")
 	}
 	if _, ok := schema.Env[schema.Root]; !ok {
-		return Schema{}, schemaError("$", CodeSchemaUnknownType, fmt.Sprintf("root %q does not resolve to a defined record", schema.Root))
+		return omnist.Schema{}, schemaError("$", omnist.CodeSchemaUnknownType, fmt.Sprintf("root %q does not resolve to a defined record", schema.Root))
 	}
 
 	// S-6: every Ref type (forward references and mutual recursion both
@@ -121,11 +125,11 @@ func (p *osdParser) parseSchema() (Schema, error) {
 	for _, name := range recordOrder {
 		rec := schema.Env[name]
 		for _, f := range rec.Fields {
-			if f.Type.Kind != TypeRefKind {
+			if f.Type.Kind != omnist.TypeRefKind {
 				continue
 			}
 			if _, ok := schema.Env[f.Type.RefName]; !ok {
-				return Schema{}, schemaError(rec.Name+"."+f.Label, CodeSchemaUnknownType, fmt.Sprintf("unknown type %q", f.Type.RefName))
+				return omnist.Schema{}, schemaError(rec.Name+"."+f.Label, omnist.CodeSchemaUnknownType, fmt.Sprintf("unknown type %q", f.Type.RefName))
 			}
 		}
 	}
@@ -135,23 +139,23 @@ func (p *osdParser) parseSchema() (Schema, error) {
 }
 
 // scalarKeyword reports whether name is one of the seven scalar-name
-// keywords (spec §5.6) and, if so, which ScalarKind it names.
-func scalarKeyword(name string) (ScalarKind, bool) {
+// keywords (spec §5.6) and, if so, which omnist.ScalarKind it names.
+func scalarKeyword(name string) (omnist.ScalarKind, bool) {
 	switch name {
 	case "string":
-		return KindString, true
+		return omnist.KindString, true
 	case "integer":
-		return KindInteger, true
+		return omnist.KindInteger, true
 	case "number":
-		return KindNumber, true
+		return omnist.KindNumber, true
 	case "boolean":
-		return KindBoolean, true
+		return omnist.KindBoolean, true
 	case "date":
-		return KindDate, true
+		return omnist.KindDate, true
 	case "time":
-		return KindTime, true
+		return omnist.KindTime, true
 	case "datetime":
-		return KindDateTime, true
+		return omnist.KindDateTime, true
 	}
 	return 0, false
 }
@@ -160,12 +164,12 @@ func scalarKeyword(name string) (ScalarKind, bool) {
 // *( "," field ) [ "," ] ] "}"` (spec §5.4), enforcing S-3 (reserved
 // names) as soon as the name is read and S-5 (unique field labels) as
 // each field is added.
-func (p *osdParser) parseRecordDef() (*Record, error) {
+func (p *osdParser) parseRecordDef() (*omnist.Record, error) {
 	if err := p.advance(); err != nil { // consume 'record'
 		return nil, err
 	}
 	if p.cur.kind != osdTokName {
-		return nil, p.errAt(p.cur, CodeParseUnexpectedToken, "expected a record name")
+		return nil, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected a record name")
 	}
 	name := p.cur.text
 	if err := p.advance(); err != nil {
@@ -173,24 +177,24 @@ func (p *osdParser) parseRecordDef() (*Record, error) {
 	}
 
 	if _, isScalar := scalarKeyword(name); isScalar {
-		return nil, schemaError(name, CodeSchemaReservedName, fmt.Sprintf("record name %q is a reserved scalar keyword", name))
+		return nil, schemaError(name, omnist.CodeSchemaReservedName, fmt.Sprintf("record name %q is a reserved scalar keyword", name))
 	}
 	if name == "any" {
-		return nil, schemaError(name, CodeSchemaReservedName, "record name \"any\" is reserved")
+		return nil, schemaError(name, omnist.CodeSchemaReservedName, "record name \"any\" is reserved")
 	}
 
 	if p.cur.kind != osdTokLBrace {
-		return nil, p.errAt(p.cur, CodeParseUnexpectedToken, "expected '{' after record name")
+		return nil, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected '{' after record name")
 	}
 	if err := p.advance(); err != nil { // consume '{'
 		return nil, err
 	}
 
-	rec := &Record{Name: name}
+	rec := &omnist.Record{Name: name}
 	seenLabels := map[string]bool{}
 	for p.cur.kind != osdTokRBrace {
 		if p.cur.kind == osdTokEOF {
-			return nil, p.errAt(p.cur, CodeParseUnexpectedToken, "unexpected end of input, expected '}'")
+			return nil, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "unexpected end of input, expected '}'")
 		}
 		field, err := p.parseField(name)
 		if err != nil {
@@ -198,11 +202,11 @@ func (p *osdParser) parseRecordDef() (*Record, error) {
 		}
 		if seenLabels[field.Label] {
 			// Path is the record itself (RecordName), not
-			// RecordName.field, per §8.4's Schema-path form: this is a
+			// RecordName.field, per §8.4's omnist.Schema-path form: this is a
 			// record-level diagnostic (the record has two fields with the
 			// same label), not a field-level one — there's no single field
 			// it names, since it's the duplication itself that's wrong.
-			return nil, schemaError(name, CodeSchemaDuplicateField, fmt.Sprintf("field %q is already defined in record %q", field.Label, name))
+			return nil, schemaError(name, omnist.CodeSchemaDuplicateField, fmt.Sprintf("field %q is already defined in record %q", field.Label, name))
 		}
 		seenLabels[field.Label] = true
 		rec.Fields = append(rec.Fields, field)
@@ -215,7 +219,7 @@ func (p *osdParser) parseRecordDef() (*Record, error) {
 		case osdTokRBrace:
 			// loop exits on the next condition check
 		default:
-			return nil, p.errAt(p.cur, CodeParseUnexpectedToken, "expected ',' or '}' after field")
+			return nil, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected ',' or '}' after field")
 		}
 	}
 	if err := p.advance(); err != nil { // consume '}'
@@ -230,7 +234,7 @@ func (p *osdParser) parseRootDef() (string, error) {
 		return "", err
 	}
 	if p.cur.kind != osdTokName {
-		return "", p.errAt(p.cur, CodeParseUnexpectedToken, "expected a record name after 'root'")
+		return "", p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected a record name after 'root'")
 	}
 	name := p.cur.text
 	if err := p.advance(); err != nil {
@@ -241,44 +245,44 @@ func (p *osdParser) parseRootDef() (string, error) {
 
 // parseField implements `field = string [ cardinality ] ":" type` (spec
 // §5.4). recordName is the enclosing record's name, used to build the
-// Schema-shaped path of any diagnostic raised while parsing this field.
+// omnist.Schema-shaped path of any diagnostic raised while parsing this field.
 //
 // Per the quoting rule (§5.2), the label MUST be a quoted string; an
 // unquoted name here is specifically schema.unquoted-label (spec §5.10's
 // `record R{a:string}` worked example), not a generic syntax error.
-func (p *osdParser) parseField(recordName string) (Field, error) {
+func (p *osdParser) parseField(recordName string) (omnist.Field, error) {
 	if p.cur.kind != osdTokString {
 		if p.cur.kind == osdTokName {
-			return Field{}, schemaError(recordName, CodeSchemaUnquotedLabel, "expected a quoted field name")
+			return omnist.Field{}, schemaError(recordName, omnist.CodeSchemaUnquotedLabel, "expected a quoted field name")
 		}
-		return Field{}, p.errAt(p.cur, CodeParseUnexpectedToken, "expected a quoted field name")
+		return omnist.Field{}, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected a quoted field name")
 	}
 	label := p.cur.strVal
 	if err := p.advance(); err != nil {
-		return Field{}, err
+		return omnist.Field{}, err
 	}
 
-	card := DefaultCardinality()
+	card := omnist.DefaultCardinality()
 	if p.cur.kind == osdTokLBracket {
 		c, err := p.parseCardinality(recordName, label)
 		if err != nil {
-			return Field{}, err
+			return omnist.Field{}, err
 		}
 		card = c
 	}
 
 	if p.cur.kind != osdTokColon {
-		return Field{}, p.errAt(p.cur, CodeParseUnexpectedToken, "expected ':' after field label/cardinality")
+		return omnist.Field{}, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected ':' after field label/cardinality")
 	}
 	if err := p.advance(); err != nil {
-		return Field{}, err
+		return omnist.Field{}, err
 	}
 
 	typ, err := p.parseType(recordName, label)
 	if err != nil {
-		return Field{}, err
+		return omnist.Field{}, err
 	}
-	return Field{Label: label, Type: typ, Cardinality: card}, nil
+	return omnist.Field{Label: label, Type: typ, Cardinality: card}, nil
 }
 
 // parseCardinality implements `cardinality = "[" ( int [ "," [ int ] ] /
@@ -286,91 +290,91 @@ func (p *osdParser) parseField(recordName string) (Field, error) {
 // grammar: whole-number bounds (via parseCardinalityBound), non-negative
 // minimum, and max >= min. path names the field this cardinality belongs
 // to, for any schema.* diagnostic raised here.
-func (p *osdParser) parseCardinality(recordName, label string) (Cardinality, error) {
+func (p *osdParser) parseCardinality(recordName, label string) (omnist.Cardinality, error) {
 	path := recordName + "." + label
 
 	if err := p.advance(); err != nil { // consume '['
-		return Cardinality{}, err
+		return omnist.Cardinality{}, err
 	}
 
 	if p.cur.kind == osdTokRBracket {
 		if err := p.advance(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
-		return Cardinality{}, schemaError(path, CodeSchemaEmptyCardinality, "cardinality must not be empty")
+		return omnist.Cardinality{}, schemaError(path, omnist.CodeSchemaEmptyCardinality, "cardinality must not be empty")
 	}
 
 	if p.cur.kind == osdTokComma {
 		// "," [ int ] "]"  ->  min = 0
 		if err := p.advance(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if p.cur.kind == osdTokRBracket {
 			if err := p.advance(); err != nil {
-				return Cardinality{}, err
+				return omnist.Cardinality{}, err
 			}
-			return Cardinality{Min: 0, Unbounded: true}, nil
+			return omnist.Cardinality{Min: 0, Unbounded: true}, nil
 		}
 		maxV, err := p.parseCardinalityBound(path)
 		if err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if err := p.expectRBracket(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if maxV < 0 {
-			return Cardinality{}, schemaError(path, CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
+			return omnist.Cardinality{}, schemaError(path, omnist.CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
 		}
-		return Cardinality{Min: 0, Max: uint64(maxV)}, nil
+		return omnist.Cardinality{Min: 0, Max: uint64(maxV)}, nil
 	}
 
 	// int [ "," [ int ] ] "]"
 	firstV, err := p.parseCardinalityBound(path)
 	if err != nil {
-		return Cardinality{}, err
+		return omnist.Cardinality{}, err
 	}
 
 	switch p.cur.kind {
 	case osdTokRBracket:
 		if err := p.advance(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if firstV < 0 {
-			return Cardinality{}, schemaError(path, CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
+			return omnist.Cardinality{}, schemaError(path, omnist.CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
 		}
-		return Cardinality{Min: uint64(firstV), Max: uint64(firstV)}, nil
+		return omnist.Cardinality{Min: uint64(firstV), Max: uint64(firstV)}, nil
 	case osdTokComma:
 		if err := p.advance(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if p.cur.kind == osdTokRBracket {
 			if err := p.advance(); err != nil {
-				return Cardinality{}, err
+				return omnist.Cardinality{}, err
 			}
 			if firstV < 0 {
-				return Cardinality{}, schemaError(path, CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
+				return omnist.Cardinality{}, schemaError(path, omnist.CodeSchemaInvalidCardinality, "cardinality bound must be non-negative")
 			}
-			return Cardinality{Min: uint64(firstV), Unbounded: true}, nil
+			return omnist.Cardinality{Min: uint64(firstV), Unbounded: true}, nil
 		}
 		secondV, err := p.parseCardinalityBound(path)
 		if err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if err := p.expectRBracket(); err != nil {
-			return Cardinality{}, err
+			return omnist.Cardinality{}, err
 		}
 		if firstV < 0 || secondV < 0 || secondV < firstV {
-			return Cardinality{}, schemaError(path, CodeSchemaInvalidCardinality, "invalid cardinality range")
+			return omnist.Cardinality{}, schemaError(path, omnist.CodeSchemaInvalidCardinality, "invalid cardinality range")
 		}
-		return Cardinality{Min: uint64(firstV), Max: uint64(secondV)}, nil
+		return omnist.Cardinality{Min: uint64(firstV), Max: uint64(secondV)}, nil
 	default:
-		return Cardinality{}, p.errAt(p.cur, CodeParseUnexpectedToken, "expected ',' or ']' in cardinality")
+		return omnist.Cardinality{}, p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected ',' or ']' in cardinality")
 	}
 }
 
-func (p *osdParser) expectRBracket() *ParseError {
+func (p *osdParser) expectRBracket() *omnist.ParseError {
 	if p.cur.kind != osdTokRBracket {
-		return p.errAt(p.cur, CodeParseUnexpectedToken, "expected ']' after cardinality")
+		return p.errAt(p.cur, omnist.CodeParseUnexpectedToken, "expected ']' after cardinality")
 	}
 	return p.advance()
 }
@@ -392,9 +396,9 @@ func (p *osdParser) parseCardinalityBound(path string) (int64, error) {
 		if err := p.advance(); err != nil {
 			return 0, err
 		}
-		return 0, schemaError(path, CodeSchemaNonIntegerCardinality, "cardinality bound must be a whole number")
+		return 0, schemaError(path, omnist.CodeSchemaNonIntegerCardinality, "cardinality bound must be a whole number")
 	default:
-		return 0, p.errAt(t, CodeParseUnexpectedToken, "expected a cardinality bound")
+		return 0, p.errAt(t, omnist.CodeParseUnexpectedToken, "expected a cardinality bound")
 	}
 }
 
@@ -403,7 +407,7 @@ func (p *osdParser) parseCardinalityBound(path string) (int64, error) {
 // diagnostic raised here (nullable-any, nullable-ref). Reference
 // resolution (S-6) is deferred to parseSchema, once every record-def has
 // been seen, since forward references and mutual recursion are legal.
-func (p *osdParser) parseType(recordName, label string) (Type, error) {
+func (p *osdParser) parseType(recordName, label string) (omnist.Type, error) {
 	t := p.cur
 	if t.kind != osdTokName {
 		// Per the quoting rule (§5.2), a quoted string in type position is
@@ -412,13 +416,13 @@ func (p *osdParser) parseType(recordName, label string) (Type, error) {
 		// via omnist-spec#35) is the dedicated code for it; any other
 		// non-name token in type position is still a generic syntax error.
 		if t.kind == osdTokString {
-			return Type{}, schemaError(recordName, CodeSchemaQuotedType, "expected a bare type name, not a quoted string")
+			return omnist.Type{}, schemaError(recordName, omnist.CodeSchemaQuotedType, "expected a bare type name, not a quoted string")
 		}
-		return Type{}, p.errAt(t, CodeParseUnexpectedToken, "expected a type name")
+		return omnist.Type{}, p.errAt(t, omnist.CodeParseUnexpectedToken, "expected a type name")
 	}
 	name := t.text
 	if err := p.advance(); err != nil {
-		return Type{}, err
+		return omnist.Type{}, err
 	}
 	path := recordName + "." + label
 
@@ -427,24 +431,24 @@ func (p *osdParser) parseType(recordName, label string) (Type, error) {
 		if p.cur.kind == osdTokQuestion {
 			nullable = true
 			if err := p.advance(); err != nil {
-				return Type{}, err
+				return omnist.Type{}, err
 			}
 		}
-		return ScalarType(kind, nullable), nil
+		return omnist.ScalarType(kind, nullable), nil
 	}
 
 	if name == "any" {
 		if p.cur.kind == osdTokQuestion {
-			return Type{}, schemaError(path, CodeSchemaNullableAny, "`any` already includes null; `any?` is not valid")
+			return omnist.Type{}, schemaError(path, omnist.CodeSchemaNullableAny, "`any` already includes null; `any?` is not valid")
 		}
-		return AnyType(), nil
+		return omnist.AnyType(), nil
 	}
 
 	// Reference: any other name (spec §5.6), including capitalized "Any",
 	// which is deliberately not the any-type — only the exact lowercase
 	// spelling is (spec §5.6's own worked note).
 	if p.cur.kind == osdTokQuestion {
-		return Type{}, schemaError(path, CodeSchemaNullableRef, "'?' cannot apply to a reference; use cardinality [0,1] instead")
+		return omnist.Type{}, schemaError(path, omnist.CodeSchemaNullableRef, "'?' cannot apply to a reference; use cardinality [0,1] instead")
 	}
-	return RefType(name), nil
+	return omnist.RefType(name), nil
 }
