+++ referee_coverage_test.go
@@ -1,6 +1,4 @@
-package omnist
-
-import "testing"
+package omnist_test
 
 // This file closes the remaining branch coverage in referee.go that the
 // 10-case spec self-test (referee_test.go) doesn't happen to exercise --
@@ -8,14 +6,22 @@ import "testing"
 // not conformance-harness execution plumbing, so it holds the repo's
 // normal 100%-per-function coverage bar rather than tools/conformance/**'s
 // documented exception (see .github/workflows/ci.yml).
+//
+// External "omnist_test" package: see referee_test.go's comment for why.
+
+import (
+	"testing"
+
+	omnist "github.com/omnist-dev/omnist-go"
+)
 
 func TestRefereeCoverageDocumentsEqualNodeVsValueMismatch(t *testing.T) {
 	node := mustOML(t, "a: 1")
 	value := mustOML(t, "42")
-	if DocumentsEqual(node, value) {
+	if omnist.DocumentsEqual(node, value) {
 		t.Fatal("a node-shaped and a value-shaped Document must not compare equal")
 	}
-	if DocumentsEqual(value, node) {
+	if omnist.DocumentsEqual(value, node) {
 		t.Fatal("comparison must be symmetric")
 	}
 }
@@ -23,7 +29,7 @@ func TestRefereeCoverageDocumentsEqualNodeVsValueMismatch(t *testing.T) {
 func TestRefereeCoverageValuesEqualNullMismatch(t *testing.T) {
 	null := mustOML(t, "null")
 	nonNull := mustOML(t, "1")
-	if DocumentsEqual(null, nonNull) {
+	if omnist.DocumentsEqual(null, nonNull) {
 		t.Fatal("null and non-null bare values must not compare equal")
 	}
 }
@@ -31,7 +37,7 @@ func TestRefereeCoverageValuesEqualNullMismatch(t *testing.T) {
 func TestRefereeCoverageValuesEqualBothNull(t *testing.T) {
 	a := mustOML(t, "null")
 	b := mustOML(t, "null")
-	if !DocumentsEqual(a, b) {
+	if !omnist.DocumentsEqual(a, b) {
 		t.Fatal("two null bare values must compare equal")
 	}
 }
@@ -39,7 +45,7 @@ func TestRefereeCoverageValuesEqualBothNull(t *testing.T) {
 func TestRefereeCoverageScalarsEqualBothNaN(t *testing.T) {
 	a := mustOML(t, "nan")
 	b := mustOML(t, "nan")
-	if !DocumentsEqual(a, b) {
+	if !omnist.DocumentsEqual(a, b) {
 		t.Fatal("two NaN-valued number scalars must compare equal under the referee (unlike plain Scalar.Equal / Go's ==)")
 	}
 }
@@ -47,7 +53,7 @@ func TestRefereeCoverageScalarsEqualBothNaN(t *testing.T) {
 func TestRefereeCoverageScalarsEqualKindMismatch(t *testing.T) {
 	integer := mustOML(t, "1")
 	number := mustOML(t, "1.0")
-	if DocumentsEqual(integer, number) {
+	if omnist.DocumentsEqual(integer, number) {
 		t.Fatal("an integer-kind and a number-kind scalar of the same magnitude must not compare equal (spec D-5)")
 	}
 }
@@ -55,7 +61,7 @@ func TestRefereeCoverageScalarsEqualKindMismatch(t *testing.T) {
 func TestRefereeCoverageTargetsEqualNodeVsValueMismatch(t *testing.T) {
 	a := mustOML(t, "m: {}")
 	b := mustOML(t, "m: 1")
-	if DocumentsEqual(a, b) {
+	if omnist.DocumentsEqual(a, b) {
 		t.Fatal("an edge whose target is a node must not compare equal to one whose target is a value")
 	}
 }
@@ -63,7 +69,7 @@ func TestRefereeCoverageTargetsEqualNodeVsValueMismatch(t *testing.T) {
 func TestRefereeCoverageNodesEqualLengthMismatch(t *testing.T) {
 	a := mustOML(t, "a: 1")
 	b := mustOML(t, "a: 1; b: 2")
-	if DocumentsEqual(a, b) {
+	if omnist.DocumentsEqual(a, b) {
 		t.Fatal("nodes with a different edge count must not compare equal")
 	}
 }
@@ -71,7 +77,7 @@ func TestRefereeCoverageNodesEqualLengthMismatch(t *testing.T) {
 func TestRefereeCoverageNodesEqualLabelMismatch(t *testing.T) {
 	a := mustOML(t, "a: 1")
 	b := mustOML(t, "b: 1")
-	if DocumentsEqual(a, b) {
+	if omnist.DocumentsEqual(a, b) {
 		t.Fatal("nodes whose edges have different labels must not compare equal")
 	}
 }
@@ -79,7 +85,7 @@ func TestRefereeCoverageNodesEqualLabelMismatch(t *testing.T) {
 func TestRefereeCoverageSchemaExactEqualRootMismatch(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } root R`)
 	b := mustOSD(t, `record R { "x": string } record S { "x": string } root S`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("schemas with different roots must not compare equal")
 	}
 }
@@ -87,7 +93,7 @@ func TestRefereeCoverageSchemaExactEqualRootMismatch(t *testing.T) {
 func TestRefereeCoverageSchemaExactEqualEnvLengthMismatch(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } root R`)
 	b := mustOSD(t, `record R { "x": string } record S { "y": string } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("schemas with a different number of records must not compare equal")
 	}
 }
@@ -95,7 +101,7 @@ func TestRefereeCoverageSchemaExactEqualEnvLengthMismatch(t *testing.T) {
 func TestRefereeCoverageSchemaExactEqualMissingRecordInOther(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } record S { "y": string } root R`)
 	b := mustOSD(t, `record R { "x": string } record T { "y": string } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("a record name present in one env but absent from the other must not compare equal")
 	}
 }
@@ -103,7 +109,7 @@ func TestRefereeCoverageSchemaExactEqualMissingRecordInOther(t *testing.T) {
 func TestRefereeCoverageRecordFieldSetEqualExactFieldCountMismatch(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } root R`)
 	b := mustOSD(t, `record R { "x": string, "y": string } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("records with a different field count must not compare equal")
 	}
 }
@@ -111,7 +117,7 @@ func TestRefereeCoverageRecordFieldSetEqualExactFieldCountMismatch(t *testing.T)
 func TestRefereeCoverageRecordFieldSetEqualExactMissingLabel(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } root R`)
 	b := mustOSD(t, `record R { "y": string } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("records whose fields have different labels must not compare equal")
 	}
 }
@@ -119,7 +125,7 @@ func TestRefereeCoverageRecordFieldSetEqualExactMissingLabel(t *testing.T) {
 func TestRefereeCoverageTypesEqualExactKindMismatch(t *testing.T) {
 	a := mustOSD(t, `record R { "x": string } root R`)
 	b := mustOSD(t, `record R { "x": R } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("a scalar-typed field and a ref-typed field must not compare equal")
 	}
 }
@@ -127,7 +133,7 @@ func TestRefereeCoverageTypesEqualExactKindMismatch(t *testing.T) {
 func TestRefereeCoverageTypesEqualExactRefNameMismatch(t *testing.T) {
 	a := mustOSD(t, `record R { "x": S } record S { "y": string } root R`)
 	b := mustOSD(t, `record R { "x": T } record T { "y": string } root R`)
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("ref fields pointing at differently-named records must not compare equal under exact mode (no renaming permitted)")
 	}
 }
@@ -135,7 +141,7 @@ func TestRefereeCoverageTypesEqualExactRefNameMismatch(t *testing.T) {
 func TestRefereeCoverageTypesEqualExactAnyKind(t *testing.T) {
 	a := mustOSD(t, `record R { "x": any } root R`)
 	b := mustOSD(t, `record R { "x": any } root R`)
-	if !SchemasEqual(a, b, ModeExact) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("two any-typed fields must compare equal")
 	}
 }
@@ -163,26 +169,26 @@ func TestRefereeCoverageIsomorphicBNameAlreadyClaimed(t *testing.T) {
 		record Root { "p": Same, "q": Same }
 		root Root
 	`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("Left and Right cannot both isomorphically map to the same B record Same")
 	}
 }
 
 func TestRefereeCoverageIsomorphicDanglingRef(t *testing.T) {
-	// Not constructible through ReadOSD (S-6 rejects a dangling ref), so
-	// build the Schema directly to exercise recordsIsomorphic's
+	// Not constructible through osd.Read (S-6 rejects a dangling ref), so
+	// build the omnist.Schema directly to exercise recordsIsomorphic's
 	// !okA || !okB branch.
-	a := Schema{
+	a := omnist.Schema{
 		Root:     "R",
 		EnvOrder: []string{"R"},
-		Env: map[string]*Record{
-			"R": {Name: "R", Fields: []Field{
-				{Label: "x", Type: Type{Kind: TypeRefKind, RefName: "Missing"}, Cardinality: Cardinality{Min: 1, Max: 1}},
+		Env: map[string]*omnist.Record{
+			"R": {Name: "R", Fields: []omnist.Field{
+				{Label: "x", Type: omnist.Type{Kind: omnist.TypeRefKind, RefName: "Missing"}, Cardinality: omnist.Cardinality{Min: 1, Max: 1}},
 			}},
 		},
 	}
 	b := mustOSD(t, `record R { "x": S } record S { "y": string } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("a dangling ref on one side must not compare isomorphic")
 	}
 }
@@ -190,7 +196,7 @@ func TestRefereeCoverageIsomorphicDanglingRef(t *testing.T) {
 func TestRefereeCoverageIsomorphicFieldCountMismatch(t *testing.T) {
 	a := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
 	b := mustOSD(t, `record S { "x": string, "y": string } record R { "a": S } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("records with a different field count must not compare isomorphic")
 	}
 }
@@ -198,7 +204,7 @@ func TestRefereeCoverageIsomorphicFieldCountMismatch(t *testing.T) {
 func TestRefereeCoverageIsomorphicMissingLabel(t *testing.T) {
 	a := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
 	b := mustOSD(t, `record S { "y": string } record R { "a": S } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("records whose fields have different labels must not compare isomorphic")
 	}
 }
@@ -206,7 +212,7 @@ func TestRefereeCoverageIsomorphicMissingLabel(t *testing.T) {
 func TestRefereeCoverageIsomorphicCardinalityMismatch(t *testing.T) {
 	a := mustOSD(t, `record S { "x" [0,1]: string } record R { "a": S } root R`)
 	b := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("differing cardinality on a matched field must not compare isomorphic")
 	}
 }
@@ -214,7 +220,7 @@ func TestRefereeCoverageIsomorphicCardinalityMismatch(t *testing.T) {
 func TestRefereeCoverageIsomorphicTypeKindMismatch(t *testing.T) {
 	a := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
 	b := mustOSD(t, `record T { "x": T } record R { "a": T } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("a scalar-typed field and a ref-typed field must not compare isomorphic")
 	}
 }
@@ -222,7 +228,7 @@ func TestRefereeCoverageIsomorphicTypeKindMismatch(t *testing.T) {
 func TestRefereeCoverageIsomorphicScalarKindMismatch(t *testing.T) {
 	a := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
 	b := mustOSD(t, `record S { "x": integer } record R { "a": S } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("differing scalar kind on a matched field must not compare isomorphic")
 	}
 }
@@ -230,7 +236,7 @@ func TestRefereeCoverageIsomorphicScalarKindMismatch(t *testing.T) {
 func TestRefereeCoverageIsomorphicNullableMismatch(t *testing.T) {
 	a := mustOSD(t, `record S { "x": string } record R { "a": S } root R`)
 	b := mustOSD(t, `record S { "x": string? } record R { "a": S } root R`)
-	if SchemasEqual(a, b, ModeIsomorphic) {
+	if omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("differing nullability on a matched field must not compare isomorphic")
 	}
 }
@@ -242,7 +248,7 @@ func TestRefereeCoverageIsomorphicNullableMismatch(t *testing.T) {
 func TestRefereeCoverageIsomorphicRevisitsAlreadyMappedName(t *testing.T) {
 	a := mustOSD(t, `record R { "a": R, "b": R } root R`)
 	b := mustOSD(t, `record S { "a": S, "b": S } root S`)
-	if !SchemasEqual(a, b, ModeIsomorphic) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("a self-referential record reached twice must still compare isomorphic once R<->S is committed")
 	}
 }
@@ -250,7 +256,7 @@ func TestRefereeCoverageIsomorphicRevisitsAlreadyMappedName(t *testing.T) {
 func TestRefereeCoverageIsomorphicAnyKind(t *testing.T) {
 	a := mustOSD(t, `record R { "x": any } root R`)
 	b := mustOSD(t, `record R { "x": any } root R`)
-	if !SchemasEqual(a, b, ModeIsomorphic) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("two any-typed fields must compare isomorphic")
 	}
 }
