+++ referee_test.go
@@ -1,6 +1,4 @@
-package omnist
-
-import "testing"
+package omnist_test
 
 // This file hand-ports the spec's own 10-case referee self-test
 // (vendor/omnist-spec/conformance/fixtures/_referee-self-test/) as real Go
@@ -8,33 +6,29 @@ import "testing"
 // before it judges anything ... all three existing ports did this as
 // their literal step one." Each test's name and comment quote the
 // fixture's own purpose.txt. This MUST pass before tools/conformance's
-// vector runner is trusted to judge anything using SchemasEqual/
-// DocumentsEqual.
+// vector runner is trusted to judge anything using omnist.SchemasEqual/
+// omnist.DocumentsEqual.
+//
+// This file (and referee_coverage_test.go) lives in the external
+// "omnist_test" package, not "omnist" -- it only ever needed exported API
+// (omnist.SchemasEqual/omnist.DocumentsEqual/omnist.ModeExact/omnist.ModeIsomorphic), so moving it out
+// avoids the import cycle package osd would otherwise create (osd imports
+// omnist; an internal omnist test file importing osd is not allowed).
+// mustOSD/mustOML are defined once, for both files, in
+// osd_external_test_helpers_test.go.
 
-func mustOSD(t *testing.T, text string) Schema {
-	t.Helper()
-	s, err := ReadOSD(text)
-	if err != nil {
-		t.Fatalf("ReadOSD failed: %v", err)
-	}
-	return s
-}
+import (
+	"testing"
 
-func mustOML(t *testing.T, text string) Document {
-	t.Helper()
-	d, err := ReadOML(text, DefaultLimits())
-	if err != nil {
-		t.Fatalf("ReadOML failed: %v", err)
-	}
-	return d
-}
+	omnist "github.com/omnist-dev/omnist-go"
+)
 
 // Case 01: fields are an unordered set at the model layer (§3.1) --
 // declaration order must not affect equality.
 func TestReferee01SchemaExactEqualDifferentFieldOrder(t *testing.T) {
 	a := mustOSD(t, "record R {\n    \"x\": string,\n    \"y\": integer,\n}\nroot R\n")
 	b := mustOSD(t, "record R {\n    \"y\": integer,\n    \"x\": string,\n}\nroot R\n")
-	if !SchemasEqual(a, b, ModeExact) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want equal (field order must not affect exact equality)")
 	}
 }
@@ -44,7 +38,7 @@ func TestReferee01SchemaExactEqualDifferentFieldOrder(t *testing.T) {
 func TestReferee02SchemaExactNotEqualCardinalityDiff(t *testing.T) {
 	a := mustOSD(t, "record R {\n    \"x\": string,\n}\nroot R\n")
 	b := mustOSD(t, "record R {\n    \"x\" [0,1]: string,\n}\nroot R\n")
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want not-equal (cardinality differs)")
 	}
 }
@@ -54,7 +48,7 @@ func TestReferee02SchemaExactNotEqualCardinalityDiff(t *testing.T) {
 func TestReferee03SchemaIsomorphicEqualRenamedRecords(t *testing.T) {
 	a := mustOSD(t, "record A {\n    \"x\": string,\n}\nroot A\n")
 	b := mustOSD(t, "record B {\n    \"x\": string,\n}\nroot B\n")
-	if !SchemasEqual(a, b, ModeIsomorphic) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeIsomorphic) {
 		t.Fatal("want equal under isomorphic mode")
 	}
 }
@@ -65,7 +59,7 @@ func TestReferee03SchemaIsomorphicEqualRenamedRecords(t *testing.T) {
 func TestReferee04SchemaExactNotEqualRenamedRecords(t *testing.T) {
 	a := mustOSD(t, "record A {\n    \"x\": string,\n}\nroot A\n")
 	b := mustOSD(t, "record B {\n    \"x\": string,\n}\nroot B\n")
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want not-equal under exact mode")
 	}
 }
@@ -76,7 +70,7 @@ func TestReferee04SchemaExactNotEqualRenamedRecords(t *testing.T) {
 func TestReferee05DocumentNotEqualReorderedEdges(t *testing.T) {
 	a := mustOML(t, "a: 1\nb: 2\n")
 	b := mustOML(t, "b: 2\na: 1\n")
-	if DocumentsEqual(a, b) {
+	if omnist.DocumentsEqual(a, b) {
 		t.Fatal("want not-equal (edge order differs)")
 	}
 }
@@ -86,7 +80,7 @@ func TestReferee05DocumentNotEqualReorderedEdges(t *testing.T) {
 func TestReferee06DocumentEqualArraySugarVsRepeatedLabels(t *testing.T) {
 	a := mustOML(t, "tag: [\"x\", \"y\"]\n")
 	b := mustOML(t, "tag: \"x\"\ntag: \"y\"\n")
-	if !DocumentsEqual(a, b) {
+	if !omnist.DocumentsEqual(a, b) {
 		t.Fatal("want equal (array sugar == repeated labels)")
 	}
 }
@@ -97,7 +91,7 @@ func TestReferee06DocumentEqualArraySugarVsRepeatedLabels(t *testing.T) {
 func TestReferee07SchemaExactEqualDifferentEnvDeclarationOrder(t *testing.T) {
 	a := mustOSD(t, "record Zeta {\n    \"x\": string,\n}\nrecord Root {\n    \"z\": Zeta,\n}\nroot Root\n")
 	b := mustOSD(t, "record Root {\n    \"z\": Zeta,\n}\nrecord Zeta {\n    \"x\": string,\n}\nroot Root\n")
-	if !SchemasEqual(a, b, ModeExact) {
+	if !omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want equal (env declaration order must not affect equality)")
 	}
 }
@@ -106,7 +100,7 @@ func TestReferee07SchemaExactEqualDifferentEnvDeclarationOrder(t *testing.T) {
 func TestReferee08SchemaExactNotEqualScalarKindDiff(t *testing.T) {
 	a := mustOSD(t, "record R {\n    \"x\": string,\n}\nroot R\n")
 	b := mustOSD(t, "record R {\n    \"x\": integer,\n}\nroot R\n")
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want not-equal (scalar kind differs)")
 	}
 }
@@ -116,7 +110,7 @@ func TestReferee08SchemaExactNotEqualScalarKindDiff(t *testing.T) {
 func TestReferee09SchemaExactNotEqualNullableDiff(t *testing.T) {
 	a := mustOSD(t, "record R {\n    \"x\": string,\n}\nroot R\n")
 	b := mustOSD(t, "record R {\n    \"x\": string?,\n}\nroot R\n")
-	if SchemasEqual(a, b, ModeExact) {
+	if omnist.SchemasEqual(a, b, omnist.ModeExact) {
 		t.Fatal("want not-equal (nullable flag differs)")
 	}
 }
@@ -128,7 +122,7 @@ func TestReferee10DocumentEqualIdenticalNestedStructure(t *testing.T) {
 	text := "order: {\n    id: \"A1\"\n    item: { sku: \"W\"; qty: 3 }\n    item: { sku: \"G\"; qty: 1 }\n}\n"
 	a := mustOML(t, text)
 	b := mustOML(t, text)
-	if !DocumentsEqual(a, b) {
+	if !omnist.DocumentsEqual(a, b) {
 		t.Fatal("want equal (identical nested structure)")
 	}
 }
